Moved FeeMode code into its own file
