+++ .scalafmt.conf
@@ -1,6 +1,6 @@
 version = "2.0.1"
 # See Documentation at https://scalameta.org/scalafmt/#Configuration
-trailingCommas = never
+# trailingCommas = never
 maxColumn = 80
 docstrings = ScalaDoc
 continuationIndent {
@@ -14,10 +14,7 @@ align {
     openParenCallSite = true
 }
 
-danglingParentheses {
-    callSite = false    
-    defnSite = false
-}
+danglingParentheses = false
 
 newlines {
     alwaysBeforeTopLevelStatements = true
